% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSpatiallyVariableFeaturesBayes.R
\name{findSpatiallyVariableFeaturesBayes}
\alias{findSpatiallyVariableFeaturesBayes}
\title{Identify spatially variable genes in a Bayesian manner.}
\usage{
findSpatiallyVariableFeaturesBayes(
  spatial.obj = NULL,
  n.spots.subsample = 500L,
  iter.per.chain = 1000L,
  warump.per.chain = 250L,
  n.chains = 1L,
  thin.rate = 5L,
  n.cores.chain = 1L,
  n.cores.per.chain = 4L,
  model.priors = NULL,
  algorithm = "meanfield",
  opencl.params = NULL,
  random.seed = 312,
  verbose = FALSE,
  save.model = FALSE
)
}
\arguments{
\item{spatial.obj}{An object of class \code{Seurat} containing sptial data. Defaults to NULL.}

\item{n.spots.subsample}{An integer specifying the number of spots per-gene to subsample to when performing estimation. Defaults to 500.}

\item{iter.per.chain}{An integer specifying the number of iterations per chain. Defaults to 1000L.}

\item{n.chains}{(Optional) An integer specifying the number of chains used when performing variational inference. Defaults to 1.}

\item{thin.rate}{(Optional) An integer specifying the thinning rate of the VI algorithm. Defaults to 5.}

\item{n.cores.chain}{An integer specifying the number of cores to be used when fitting the Bayesian hierarchical Gaussian process model. Defaults to 1.}

\item{n.cores.per.chain}{An integer specifying the number of cores to be used within each chain when fitting the Bayesian hierarchical model. Defaults to 4.}

\item{model.priors}{A vector containing priors to be used in model fitting. If left NULL, intelligent priors will be set internally. See \code{\link[brms]{set_prior}} for details. Defaults to NULL.}

\item{algorithm}{(Optional) A string specifying the variational inference or sampling algorithm to be used. Must be one of "meanfield", "fullrank", "laplace", "pathfinder" (all approximate methods) or "sampling" for MCMC via NUTS. Note that MCMC will be slower than using a VI algorithm or the Laplace approximation. Defaults to "meanfield".}

\item{opencl.params}{(Optional) A two-element double vector specifying the platform and device IDs of the OpenCL GPU device. Most users should specify \code{c(0, 0)}. See \code{\link[brms]{opencl}} for more details. Defaults to NULL.}

\item{random.seed}{A double specifying the random seed to be used when fitting and sampling from the model. Defaults to 312.}

\item{verbose}{(Optional) A Boolean specifying whether or not verbose model output should be printed to the console. Defaults to FALSE.}

\item{save.model}{(Optional) A Boolean specifying whether or not the fitted model generated by \code{\link[brms]{brm}} should be saved to the unstructured metadata of \code{spatial.obj}. Defaults to FALSE.}

\item{warmup.per.chain}{An integer specifying the number of warmup (burn-in) iterations per chain. Defaults to 250.}
}
\description{
This function implements HVG estimation using Bayesian variational inference to approximate the posterior distribution of the mean and overdispersion of each gene.
}
\details{
\itemize{
\item This function utilizes a multivariate Gaussian process (GP) to account for spatial variation while estimating per-gene means, variances, and dispersions.
}
}
\seealso{
\code{\link[Seurat]{FindSpatiallyVariableFeatures}}

\code{\link[brms]{brm}}
}
\author{
Jack R. Leary
}
