% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariableFeaturesBayes.R
\name{findVariableFeaturesBayes}
\alias{findVariableFeaturesBayes}
\title{Identify highly variable genes in a Bayesian manner.}
\usage{
findVariableFeaturesBayes(expr.mat = NULL)
}
\arguments{
\item{expr.mat}{Either an object of class \code{Seurat} or \code{SingleCellExperiment}. Defaults to NULL.}
}
\value{
Depending on the input, either an object of class \code{Seurat} or \code{SingleCellExperiment} with HVG metadata added, or a \code{data.frame} containing HVG results.
}
\description{
This function implements HVG identification using Bayesian inference to model the posterior distribution of the mean and variance of each gene.
}
\seealso{
\code{\link[Seurat]{FindVariableFeatures}}
}
\author{
Jack R. Leary
}
