% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMarginal.R
\name{sampleMarginal}
\alias{sampleMarginal}
\title{Sample from the marginal distribution of a fitted model.}
\usage{
sampleMarginal(marginal.dist = NULL, n = 1000L)
}
\arguments{
\item{marginal.dist}{The marginal distribution to be sampled from. Defaults to NULL.}

\item{n}{An integer specifying the number of samples. Defaults to 1000.}
}
\value{
A numeric vector of samples.
}
\description{
A simple wrapper function to generate a given number of samples from the marginal distribution of a fitted model.
}
\seealso{
\code{\link[INLA]{inla.rmarginal}}

\code{\link[INLA]{inla.is.marginal}}
}
\author{
Jack R. Leary
}
