% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariableFeaturesBayes.R
\name{findVariableFeaturesBayes}
\alias{findVariableFeaturesBayes}
\title{Identify highly variable genes in a Bayesian manner.}
\usage{
findVariableFeaturesBayes(
  sc.obj = NULL,
  subject.id = NULL,
  n.cells.subsample = 500L,
  iter.per.chain = 1000L,
  warmup.per.chain = 250L,
  n.chains = 3L,
  thin.rate = 5L,
  n.cores.chain = 3L,
  n.cores.per.chain = 2L,
  opencl.params = NULL,
  random.seed = 312,
  verbose = FALSE
)
}
\arguments{
\item{sc.obj}{An object of class \code{Seurat} or \code{SingleCellExperiment}. Defaults to NULL.}

\item{subject.id}{A string specifying the metadata column in \code{sc.obj} that contains subject IDs. Defaults to NULL.}

\item{n.cells.subsample}{An integer specifying the number of cells per-gene to subsample to when performing estimation. Defaults to 500.}

\item{iter.per.chain}{An integer specifying the number of iterations per chain. Defaults to 1000L.}

\item{warmup.per.chain}{An integer specifying the number of warmup (burn-in) iterations per chain. Defaults to 250.}

\item{n.chains}{(Optional) An integer specifying the number of chains used when performing variational inference. Defaults to 3.}

\item{thin.rate}{(Optional) An integer specifying the thinning rate of the VI algorithm. Defaults to 5.}

\item{n.cores.chain}{An integer specifying the number of cores to be used when fitting the Bayesian hierarchical model. Defaults to 3.}

\item{n.cores.per.chain}{An integer specifying the number of cores to be used within each chain when fitting the Bayesian hierarchical model. Defaults to 2.}

\item{opencl.params}{(Optional) A two-element double vector specifying the platform and device IDs of the OpenCL GPU device. Most users should specify \code{c(0, 0)}. See \code{\link[brms]{opencl}} for more details. Defaults to NULL.}

\item{random.seed}{A double specifying the random seed to be used when fitting and sampling from the model. Defaults to 312.}

\item{verbose}{(Optional) A Boolean specifying whether or not verbose model output should be printed to the console. Defaults to FALSE.}
}
\value{
Depending on the input, either an object of class \code{Seurat} or \code{SingleCellExperiment} with gene-level statistics added to the appropriate metadata slot.
}
\description{
This function implements HVG estimation using Bayesian variational inference to approximate the posterior distribution of the mean and dispersion of each gene.
}
\seealso{
\code{\link[Seurat]{FindVariableFeatures}}

\code{\link[scran]{modelGeneVar}}

\code{\link[brms]{brm}}
}
\author{
Jack R. Leary
}
