% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariableFeaturesBayes.R
\name{findVariableFeaturesBayes}
\alias{findVariableFeaturesBayes}
\title{Identify highly variable genes in a Bayesian manner.}
\usage{
findVariableFeaturesBayes(
  sc.obj = NULL,
  subject.id = NULL,
  n.cells.subsample = 500L,
  n.chains = 4L,
  thin.rate = 5L,
  n.cores = 4L,
  random.seed = 312
)
}
\arguments{
\item{sc.obj}{An object of class \code{Seurat} or \code{SingleCellExperiment}. Defaults to NULL.}

\item{subject.id}{A string specifying the metadata column in \code{expr.mat} that contains subject IDs. Defaults to NULL.}

\item{n.cells.subsample}{An integer specifying the number of cells per-gene to subsample to when performing estimation. Defaults to 500.}

\item{n.chains}{(Optional) An integer specifying the number of chains used when performing variational inference. Defaults to 4.}

\item{thin.rate}{(Optional) An integer specifying the thinning rate of the VI algorithm. Defaults to 5.}

\item{n.cores}{An integer specifying the number of cores to be used when fitting the Bayesian hierarchical model. Defaults to 4.}

\item{random.seed}{A double specifying the random seed to be used when fitting the model. Defaults to 312.}
}
\value{
Depending on the input, either an object of class \code{Seurat} or \code{SingleCellExperiment} with gene-level statistics added to the appropriate metadata slot.
}
\description{
This function implements HVG estimation using Bayesian variational inference to approximate the posterior distribution of the mean and dispersion of each gene.
}
\seealso{
\code{\link[Seurat]{FindVariableFeatures}}

\code{\link[scran]{modelGeneVar}}
}
\author{
Jack R. Leary
}
